+++ packages/e2e-test-utils/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Unreleased
 
+### Enhancement
+
+-   Added a `adminPage` parameter to `setOption` function so we can update options for other Settings pages. [#39502] (https://github.com/WordPress/gutenberg/pull/39502)
+-   Added the capability of handling checkboxes to `setOption` function. [#39502] (https://github.com/WordPress/gutenberg/pull/39502)
+-   Added a new `trashAllComments` function.
+
+
 ## 7.0.0 (2022-03-11)
 
 ### Breaking Changes